+++ admin-web/src/App.vue
@@ -1,3 +1,7 @@
 <template>
-  <router-view />
+  <router-view v-slot="{ Component, route: currentRoute }">
+    <transition name="route-fade" mode="out-in">
+      <component :is="Component" :key="currentRoute.path" />
+    </transition>
+  </router-view>
 </template>